Create 1. Checkmate
Your job is to find which queens can capture the king and print them. The moves that the queen can do is to move diagonally, horizontally and vertically (basically all the moves that all the other figures can do except from the knight). Beware that there might be queens that stand in the way of other queens and can stop them from capturing the king. For more clarification see the examples.